Moves tests for pwv data retrieval into seperate file
